[Core] Refactor the list handling code
The initial implementation was requiring list elements to be in a block
all by themselves, but also indented. This diverged from upstream SAM,
but also made it more complicated to handle the case, as when one drafts
a section, when a section is contained entirely of list elements. The
parser was handling the section block as a field definition followed by
a list.